Add cone-based tube filter
This tube filter fulfills the need for a polyline-to-tube filter. This
should be obsoleted in favor of an actual tube filter, when one comes
around.

This requires a newer version of vtk.js.
NOTE: The given vtk.js version will break picking due to a bug in
vtkImageMapper. Until that is done, please use vtk.js from the master
branch, as the issue is fixed there.