FIX: Camera not resetting on new volume
When the previous volume has hidden visibility and a new volume is
loaded, the camera doesn't reset onto the new volume because the actor's
visibility is set after the camera tries to find the new volume. This
fixes the issue by setting the actor's visibility before the new volume
is added. Also, the actor's visibility should be "visible" since an
image change triggers the ControllableVolumeView's state to reset to
"visible".